ISubscriptionOptions to internal
