+++ src/Tests/SourceGenerationContext.cs
@@ -1,0 +1,21 @@
+using System.Text.Json.Serialization;
+using PostgresOutbox.Serialization;
+
+namespace Tests;
+
+[MessageUrn("user-created:v1")]
+internal record UserCreated(
+    Guid Id,
+    string Name
+);
+
+[MessageUrn("user-deleted:v1")]
+internal record UserDeleted(
+    Guid Id,
+    string Name
+);
+
+[JsonSourceGenerationOptions(WriteIndented = true)]
+[JsonSerializable(typeof(UserCreated))]
+[JsonSerializable(typeof(UserDeleted))]
+internal partial class SourceGenerationContext: JsonSerializerContext{}
